Add a new (private) widget used for blinking buttons.
This widget implements blinking buttons (for now only the
old blinking style like in netk, glowing will come later).
This widget is used in both the task list and the window
list plugin.

The API is not public available, since I'm not sure the
naming is good and if it's useful for other plugins. So for
now it will be only used internally to reduce code
dupplication.