Add blinking support to the arrow button.
The arrow button supports blinking by now. You can create a
normal toggle button by using GTK_ARROW_NONE as arrow type.
It is also possible to add child widgets to the arrow button,
previously this was not possible (and not needed). When both
an arrow is set and a child is added, the arrow will show
shown next to the child, this will be used for group buttons in
the tasklist plugin.

Probably some allocation issues will show up, but it's a good
start for reducing the code complexity in the wnck plugins.