Add style properties for max-button-width and ellipsize-mode.
The max-button-width style property can be used to define the
maximum button size. When the value is set to -1, the buttons
will be expanded across the entire panel.

The ellipsize-mode style property can be used to set the
ellipsize mode of the tasklist button. Some people prefer
omitting characters in the middle.

Put something like this in your .gtkrc-2.0 to make it all work:

style "xfce-tasklist-style"
{
  XfceTasklist::max-button-length = -1
  XfceTasklist::ellipsize-mode = PANGO_ELLIPSIZE_MIDDLE
}
class "XfceTasklist" style "xfce-tasklist-style"