Improved panel window code.
Not entirely finished but good enough for normal usage. The code
basically splits the window layer in two: a base window class that
is used for the normal window and the auto hide window and
a panel window class.
The screen positioning code has been improved too and should work
a lot better with multi screen setups and randr.

Also dropped the last bits of the glue code.