Force internal plugins with a pre-init function to run external.
The pre-init function is only supported in the wrapper, because
panels' gtk_init() is called way before the plugins are initialized.

If an internal plugin is loaded we check if the modules has a
pre-init function, if so, print a warning and force the plugin to
run in a wrapper.