Simplify the itembar code and add support for wrapping.
Use the panel window for dnd support on the panel, this removed
a bunch of code from the itembar and makes things easier to
understand.

Also add support for wrap items in the itembar code, this
will be available in the next commit through the separator plugin.

The dnd visualisaion is also improved and works better then the
3px black line we used in previous versions of the panel; can
still be improved to moving existing items on a panel, but that
will come later.

Furthermore some related code cleanups and implement child
properties in the itembar so it is a real container.