Use a timer for a minimum time between auto plugin restart.
The restart counter was a bit flawed, because if the panel would
run for weeks and 3 restarts occured, it's not a big deal, still
the panel would've bothered the user what to do.

We now use a minimum time between restarts, if there was another
restart in the last 60 seconds, ask the user what to do.