Add new plugin directory menu.
Plugin to walk to a directory tree using a menu. Folders
are loaded on the fly and desktop files are handled
like menu items (unix only).
It is also possible to set a file pattern to allow files
in the menu (instead of only directories). The files are
opened in the default handler.