Make the systray takeover selection.
This instead of ignoring the startup. Since your plugin
is unique now, it won't give problems when one panel is
running. It also makes the systray restart after a crash
work better, since most of the time is takes a while (in
cpu cycles) before the selection is released of the crashing
plugin.