Add logout to plugin menu and drop quit and restart.
In 4.6 the quit action also showed the logout dialog, instead of
quiting the panel. Also drop the restart option since it is not
suitable for normal users (xfce4-panel -r still exists).
This makes it harder to accidentally remove the panel from the
session, if you want that, run xfce4-panel -q (normal quit).