Improve code for external plugins.
Split the base of the external plugin code in three pieces:
PanelPluginExternal as abstract object to handle the child
monitoring and access-point for the panel and
PanelPluginExternal{46,Wrapper} to handle the dbus or socket
communication. This drop a lot of duplicated code.

Also improve handling of external plugins during a child
crash; we now wait until the child is terminated before
we launch the new proccess and also handle restarts of a
child (requested by the panel) without bothering the
user.