Add debug mode to run plugins in gdb.
With PANEL_DEBUG=gdb plugin will be started in gdb
and log files (with backtrace and register dump) are
automatically stored in /tmp.
Some functionality will not be available, like
automatically restart the plugins.

Also cleanup the debug level handling a bit, so it is
easier to extend and make struts debugging work again.