Improve support for a menu editor in the applications menu.
Add support for alacarte in the applications
menu by supporting the applications.menu file and adding
an "Edit Menu" button when alacarte is installed.

Technically this is a new feature, but this is broken since 4.6
and I don't want it to be that worse in 4.8 since it is one
of the most requested features in Xfce.