Move hidden name handling to plugin.
Improve readability of the box widget by moving names
handling to the panel plugin. Use _(s,g)et_hidden in the
socket as a proxy item to store the hidden state of the icon.

Also make the socket store its name, so we don't need to
duplicate it in other widgets.