Again allocate 1x1 requests offscreen in the systray (bug #7143 and #7057).
But with 1 difference; it seems some implementations wait
(like nm-applet) on the size-changed signal. Although this
is a stupid trick, we can easily work around it by allocating
a normal size offscreen; this will trigger the signal and a new
allocation on the tray when the app completed the status icon
setup.

For dead tray icons that are still in the panel, the 1x1 hack
will continue to work since they have no window.