Don't disappear when wm restarts with compositing (bug #7194).
Because of a quick widget hide/show to make sure the
window was redrawn, the window became invisible when quickly
restarting the wm. Remove this unneeded trick.

Also cleanup the colormap handling, we now only set the rgba
colormap in the init function, since it depends on the screen
and won't change during runtime, this makes things a lot easier.