Unset visible event window in 4.6 plugins.
A lot 4.6 plugins have an event window as plugin child,
but most of them do not disable the visible event window,
resulting in an ugly looking plugin with composited panels.

So detect if the plugin has an event window as child and
hide the visible window.

For this change, plugins need to recompile.