Respawn a child if the widget was realized with a pid.
In some cases the plugin is unrealized, for example when
moving the panel to another screen. In this case there is still
a pid when realizing, while the process was asked to quit in
unrealize. In this case wait for the pid to quit and then
respawn the plugin again.