Increase size of the autohide window by 1 in each direction.
For some reason the autohide window would not pop up if I moved the
mouse to the very edge of the screen (last pixel in any direction). So
this commit increases the window size by 1 in each direction
independently of whether it is located at the top/bottom/right/left of
the screen.

I wouldn't be surprised if there is a better solution. Nick?