Set window colormap and change autohide window type (bug #7435)
The colormap was not properly applied to the autohide window
and therefore the alpha value was not rendered.

Also set the GTK_WINDOW_TOPLEVEL type on the popup windows so
the autohide window is not visible in fullscreen applications.

(cherry picked from commit 27dfa9b5bf3216ba725df55172ba861281d71aa8)