Panel: Make sure the panel has a position on startup (bug #8287).
If it happens the panel looses some settings and the position
string is also lost, it won't be visible on the next restart.
Make sure the position is set, if it is not restored, set something
new. Although this is not the position the user defined, it is
better then no panel at all.