pager: fixed wnck pager aspect ratio in deskbar mode
Previously we were relying on new api added in libwnck-2.31.
Unfortunately, stable version of libwnck was never released and the API
was not added to libwnck-3.

This commit changes the method of setting the aspect ratio.
Aspect ratio is now calculated in the plugin using screen dimensions.

One minor limitation of this method is that we are not accounting
for wnck pager frames, so the aspect ratio may be slightly inaccurate.