Improve the tasklist multimonitor handling.
The tasklist widget has an option to only show windows from the
current monitor. This previously worked by calculating the centre
pixel of the window, and then testing that against the monitor.
However, it is possible to position a window such that the centre
pixel does not appear on any monitor.

Instead, find the monitor which has the largest intersection with
the window and display the window button there.