Fix transparency with alpha being 100 (Bug 10658)
A panel with style set to 'none (use system style)' and alpha to
100 is still affected by the problem which can be triggered manually
by launching another indicator after the initial panel start. This
patch removes an unneeded check that was causing the issue. Thanks
to Thaddäus Tintenfisch for reporting the issue and finding the
exact part of the code that needed to be corrected.