Implement intelligent autohide behavior
This implementation of intelligent autohiding works as follows:

The current WnckScreen is monitored. Whenever the active window changes
or whenever the geometry of an already active window changes, the
algorithm checks whether the active window and the panel overlap. If
this is the case, the panel is hidden. Otherwise, it is made or remains
visible.

Desktop windows are treated special. Even though they overlap
with panels technically, we don't consider this relevant and always make
the panel visible when a desktop window becomes active.

Internally, there are three autohide behaviors now: never (disables
autohiding altogether), intelligent (implements the above method) and
always (the old "hide when not focused/grabbed by mouse or keyboard).