Fixing problems with intelligent panel hiding, bug #11331
ochosi:
"Just to document the findings I made during a quick debug session
with the panel from git master (4.11.x). Intelligent hiding
generally works, but the mouse-pointer behavior is buggy.
Steps to reproduce the bug:
1) set panel to intelligent hiding
2) give a window focus that doesn't overlap with the panel
3) hover the panel with the mouse
4) upon leaving the panel hides again, even though there
   is no window overlapping

What it should do instead: if intelligent hiding is in use
it should check (possibly in panel_window_leave_notify_event
or _enter_ in panel-window.c) whether there are windows
overlapping with the panel and if not, the panel shouldn't
hide on leave_notify."