Use actual output geometry to place panels
Bug: 11058

The logical screen width/height as reported by X may be larger than
the outer bounds of all the monitors combined.

Compute the actual outer bounds of the monitors and place the panel
accordingly.

But keep the struts based on the logical screen width/height as this is
what the standard expects.