Automatically migrate old "autohide" properties to "autohide-behavior"
We could write a migration script for it to avoid additional code being
added to the panel itself but the solution presented here is simple and
works: when a panel is created (e.g. during startup), we check whether
there is an old "autohide" property in the Xfconf channel. If the new
"autohide-behavior" property is not yet set, we translate "autohide"
FALSE and TRUE into "autohide-behavior" NEVER and ALWAYS, respectively.
We then reset the "autohide" property and thereby achieve an automatic,
transparent migration from old to new.

Rebased for the Gtk3 port by Simon Steinbeiss <simon@xfce.org>