Add hover and selected effect to workspace switcher
As the hover and selected state are derived of the
background color of the panel (which has no "selected"
or "active" state we could take the color from) we
still have to find a way to support dark and bright
panels out of the box.
FIXME: The current solution only works for bright themes
like Adwaita.