Prevent panel crash on unsupported GValues in PluginEvents
A non supported gvalue that slips into PluginEvent causes the panel to crash,
due to an assertion failure in the convertion between the GValue and the
GVariant. So throw a warning instead.

Example to trigger the crash
dbus-send --session --type=method_call --print-reply \
--dest='org.xfce.Panel' '/org/xfce/Panel' \
'org.xfce.Panel.PluginEvent' string:whiskermenu \
string:popup variant:byte:'a'

Unless we want the conversion to handle all possible types,
throwing just a warning instead of crashing is the good way to go.