clock: Fix invalid datetime formats appearing in list (Bug #11527)
Unfortunately some format specifiers don't seem to work in specific locales,
as it was reported with Germand and %r. Now we simply don't show these formats
in the default list of selections anymore, but they can still be set through
custom formats (thus producing an empty clock widget). We would have to add
a validator function for the gtkentry to completely fix this problem.