Workaround for xinput2's subpixel handling (LP #1795135)
This xinput2 subpixel handling results in continuous firing of
enter/leave-notify events when the panel is at 0x0 and the pointer is a
y/x=0. The actual coordinates returned by xorg can be even below 0, so
e.g. -0.67823, thereby triggering a leave-notify event.
This problem has to be tackled within xorg, clamping the subpixel
coordinates at least at the lower maxima.