launcher: Add 'New Link' button to properties (Bug #14855)
This action was previously hidden in the context menu of the treeview.
I also re-ordered the items in the properties dialog to match the order
of the items in the context menu, re-added the dialog button icons that
were lost with the GtkStock transition and made the move-up/down buttons
linked.