tasklist: Don't listen for monitor changes in the plugin
Previously both the panel and the plugin listened for monitor changes.
This lead to a race condition since there was no control over which
would receive the event first. When the panel receives the event it
may move itself to another monitor. If the plugin receives the event
it will update itself before the move, and then its contents will be
wrong after the move.

The fix makes the plugin listen to configure events from the panel
instead of monitor events from Gdk. These are only sent after the
panel has finished processing the move event. They are sent in some
other situations too, so this is slightly in-efficient, but it is
probably not worth worrying about.

Fixes bug #10725