Fix struts on the bottom and right of the screen
We can't use the monitor geometry of a single monitor to calculate
this for two reasons. Both reasons are caused because the struts
are relative to the bottom and right edges of the X11 screen.

If we try to use a single monitor geometry by saying something like
geometry.y + geometry.height, this will fail on at least one monitor
if the bottom edges of the monitors are not aligned.

If we try to calculate the rectangle containing all monitors this
will also fail because the logical X11 screen size can be even larger.

The only way to get this information is to go to raw X11 calls, or
just use the deprecated functions.

Also switch to using alloc instead of base coordinates. When this
function is called, base_x and base_y do not contain correct values.
This has no effect on the code since scale factor is the same for
all monitors under X11, and if it weren't the rest of the code would
likely break anyway.

Fixes bug #14886