Fix unreadable panel items without compositing (Bug #14842)
If the "leave-opacity" is defined then disabling compositing results
in the panel being drawn with wrong colors. The widget opacity will
somehow still be taken into consideration and results in almost
unreadable icons and text.
We work around this problem by making sure the "leave opacity" is always
100% visibility without compositing and remembering the original value
so we can restore it if compositing is re-enabled and the panel hasn't been
restarted meanwhile.