Fix anchors for gtk_menu_popup_at_widget
This fixes the use of anchors throughout the panel codebase.

In most cases the menu position depends on the panel orientation.
If the panel is vertical then the top left corner of the menu is
aligned with the top right corner of the widget. If the panel is
horizontal then the top left corner of the menu is aligned with
the bottom left corner of the widget.

In one case the logic is different: the menu which pops up when
right clicking the window list dropdown is adjusted to behave
like a normal cascading menu.

Only one usage of gtk_menu_popup_at_widget is not modified:
The dropdown menu for when a launcher has multiple actions appears
to be correct already.

Note that gdk automatically flips anchors to make the menu fit
on screen, so this automatically works if the panel is on the
right or bottom edge of the screen.

Fixes bug #14803