tasklist: Replace deprecated GdkScreen calls with GdkDisplay
The stored GdkScreen is now only used when first getting the
WnckScreen, so it can be removed. Instead, the GdkDisplay is
stored, as it is now used instead of GdkScreen in all the
other places. Additionally the monitor update function now
always keeps n_monitors updated, instead of only updating it
when all_monitors is FALSE.

gdk_display_get_n_monitors requires at least Gtk+3.22.