tasklist: Add support for larger application icons
As libwnck now supports larger icon sizes we should make use of them.
In some cases the icon-cache of libwnck is outdated. To be exact:
a correct icon is there, but the size is not. So we sometimes get a
32px icon even though we requested a larger one.
The only way to work around this is to take the icon of an app in
that group until the cache is valid again.