Revert "tasklist: Add support for larger application icons"
Unfortunately the global icon cache of libwnck lets us run into an
"icon-changed" signal loop if a user has multiple tasklists with
different icon sizes. This was already realized in 2002, when this call
was originally removed from libwnck. It's unclear to me why it was
re-introduced without addressing this issue.
For reference:
 * https://github.com/GNOME/libwnck/commit/4b216137230d2fac212b42b42acb8267e604c0a7

This reverts commit fc4efcc991b21d191f0b82412fb33070009780c5.