clock: binary: add new binary time mode
Typical binary clocks simply take the normal time and display it in
binary, this provides no value.

In binary bits should have meaning: the most significant bit should be
significant. If a day is 0x10, half a day should be 0x8.

If we divide the day in 16 equal parts that number can be represented
with four bits, noon is 0b1000 and therefore everything after noon has
the most significant bit on.

Binary time divides the day in 2 ** 16 equal parts, with 4 bits we have
a resolution of hours, and 8 bits a resolution of minutes (~6 minutes),
for seconds we need all 16 bits.

In binary time every bit is useful and it means something, now there's
actually a reason for a binary clock.

More explanation can be found by reading of hexadecimal time [1], which
is another representation of the same thing.

https://en.wikipedia.org/wiki/Hexadecimal_time

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>