Retain original GDK_CORE_DEVICE_EVENTS setting for plugins (Bug #15044)
Previously the panel disabled smooth scrolling for all plugins - and
therefore unfortunately also for all applications launched through panel
plugins - because of leave/enter notify spam when hovering the panel.
With this patch it retains the original setting and passes it down
correctly to plugins, which should in turn mean applications would get
the correct setting.