tasklist: Don't show actionsmenu in grouped windows (Bug #15622)
Previously right-clicking a window in a group (which is a menuitem) the
regular wnck actions menu would be shown and allow you to interact with
the application/window in question. Unfortunately the two menus stacked
on top of each other resulted in a bug, which caused me to 1)
investigate and 2) decide that there are no other contexts in which we
allow right-click menus in regular menus.

The new behavior is simply that right-clicks will be treated like
left-clicks and activate the window in the menu. If you want to interact
with one of the windows in the group, you can simply right-click the
group button in the panel and get the regular arrow sub-menus, which are
much more discoverable.