tasklist: Draw grouped windows count indicator (Bug #10844)
Instead of showing the number of windows in a group in as part of the
label we show it as a cairo drawn indicator on top of the button.
There are several benefits to this.
1) Also with button labels disabled, users can distinguish single
windows from grouped windows.
2) It's visually easier to parse.
3) There's less ambiguity, as some applications (e.g. Telegram) show
things like unread message as part of the label, which then looks the
same as the window count.