tasklist: Only base bg color of group indicator on gtk theme
The previous approach never worked because we cannot easily get the
*actual* background color of the panel (e.g. users can override the
background color or use an image). So we now just take the fg color,
which is reliable, as it's used for all text on the panel, use that for
drawing the bubble and then "calculate" whether to write into that with
white or black. This should work in most cases.