arrow-button: Remove timeout on dispose (Bug #15696)
The blinking source timeout has a GDestroyNotify function associated
that will update the button when the timeout is removed.

But the timeout is removed on the button finalize, i.e. once the object
is freed, which will lead to a crash trying to access memory already
freed.

I think the original author meant to use `dispose` instead of `finalize`
for removing the blinking timeout.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>