systray: Fix icons without compositing (Bug #14577)
Wrapper window uses RGBA visual when it's available;
systray manager should try to do the same for
the transparent or parent-relative background to work.

That means that the manager and icons will work with
RGBA visual even if the screen is not compositing; so
we should always force-redraw the icons.

Refs: https://bugzilla.xfce.org/show_bug.cgi?id=14577
Signed-off-by: Ivan A. Melnikov <iv@altlinux.org>