Add xfce_panel_plugin_menu_destroy to plugin API
This call allows plugins to destroy and recreate the plugin's
context menu. This is not necessary for most plugins, as the
xfce_panel_plugin_menu_insert_item refreshes the menu.
In the case of the launcher's action menu we need to more frequently
refresh the menu and also be able to remove all custom icons at once.