panel: Mitigate a serious memory leak
Sometimes, maybe most of the time, panel items are not finalized when
they are removed from the panel. For lack of being able to find the
reason why extra references accumulate, this commit eliminates them
after the item has been removed from the panel and all priority
operations have been performed, until triggering the destruction of the
item.

Beyond the beneficial effect on memory, this also avoids potential
crashes due to the reuse of data supposed to be freed, such as
typically the dbus path.