libxfce4panel: Review memory management for context menu
There should be no `g_object_unref()` in a "destroy" signal handler,
particularly when the widget destruction, and therefore the signal
emission, are triggered by a previous `g_object_unref()`. This is the
cause of a GObject-CRITICAL here, that appears for example when
removing a workspace switcher (pager plugin) via its context menu.

More generally, the memory management of custom menu items should be
simplified:
* use `g_slist_free_full()` to free the maintained items list;
* let GTK take care of its part of memory management when the menu is
  destroyed (by one means or another).

Fixes #452, see !46 for more details.