libxfce4panel: Defer context menu destruction if popped up
Since a method allowing to request the context menu destruction is
exposed in the API, care should be taken to ensure that such a request
does not occur when the menu is popped up in response to a click event.
This simple patch just defer the destruction when the menu is shown.

Fixes #442, see !45 for more details.