Use symbolic icons
Some icons in xfce4-panel were symbolic,
some not (for examples, the Help button
icon was not but the Close button icon was).
This switches nearly all of the rest of the
icons to use symbolic icons. Also adds
a little spacing between the icons and labels
on some buttons.