[FIX] menubutton stays active after selection with keyboard
This is a fix for #22.

I noticed this issue when trying to use the xfce4-popup-windowmenu script.
At first it seemed to work correctly. But i noticed that i could not trigger
the menu more than one time. After some diggin and fiddling i saw that the
button "got stuck" in toggled on, prohibitting it from being activated again.

This actually also was an issue without the xfce4-popup-windowmenu script.
It can easily be reproduced by bringing up the menu, and selecting something
with the keyboard (enter key). Now CLICK on the menubutton again, notice that
you have to click it twice (one time to deactivate it).

This MR resolves this issue by listening for the signal "selection-done", and
also sending said signal when selection is done with the keyboard.