Use a normalized URI internally for the background image
This avoids character escaping problems, for example when using the URI
in CSS code. Before this change, it was not possible to use an image
whose name contained a single quote.

It is still possible to specify a file path via Xfconf as before, and it
becomes additionally possible to specify an URI (in fact all
possibilities accepted by `g_file_new_for_commandline_arg()` to build a
GFile). So this should not cause any backward compatibility problems.