panel: Fix regression "'Span Monitor' has no effect"
Regression introduced in a24b0214c9218ab39581d0e40b1b441913d631f7,
where the string comparison with "Automatic" has been forgotten at this
place.

The feature could sometimes work despite this since then, because when
a new panel is created, the "output-name" property is left empty in
Xfconf, even though it appears as "Automatic" in the panel preferences
dialog. But as soon as the preference was changed, the feature was
broken.

Fixes #405.