systray: Check that the proxy exists when invalidating item
Previous commits have removed possibles calls to `sn_item_invalidate()`
in `sn-item.c` when the proxy has not yet been created. But there is
still a possible call from `sn-backend.c`, in response to a
"status-notifier-item-registered" signal, for which we cannot ensure
that the proxy already exists.

It is therefore better to make this check legal, rather than generating
irrelevant warnings in this case.