tasklist: Hide wireframe on window buttons
This prevents the wireframe from interfering with the receipt of pointer
events on the window buttons, which trigger its display. Without this,
an infinite loop can occur where one leaves the button to enter the
wireframe when it is displayed, causing it to be hidden and thus
re-entering the button, etc.

Fixes: #543
MR: !70