panel: Add autohide count at application level
Generally speaking, this is a notion that is missing in the code,
because the autohide count can be changed for all windows at once, or
for a particular window.

This is especially required when a new window is created from the
preferences dialog, and its autohide count needs to be set to the level
of the other windows. Otherwise, it may go negative when the dialog is
closed, triggering an xfce4-panel-CRITICAL warning.