libxfce4panel: Remove redundant `PLUGIN_FLAG_BLOCK_AUTOHIDE`
This flag is in fact not only redundant, but also conflicts with the
lock count, which are two different concepts. We keep only the lock
count, for consistency with what is done in `panel-window.c`, and
because it was already quite clear that
`xfce_panel_plugin_block_autohide()` should be used in lock/unlock
pairs.

Therefore, this commit is not pure refactoring: if external plugins were
to misuse `xfce_panel_plugin_block_autohide()`, i.e. without making sure
that a lock is followed by an unlock, it could lead to a bug. But it
would then be up to the plugin to rectify its use of
`xfce_panel_plugin_block_autohide()`.

MR: !77