Block autohide when plugins preferences dialog is shown
To apply this modification to external plugins as well, we would have to
modify the "configure-plugin" signal, adding as return value the plugin
preferences dialog. This means introducing a new signal and deprecating
the old one, without saving a modification at the plugin level to change
the signal used. So it's probably not worth it, since
blocking/unblocking autohide on showing/hiding the dialog is a matter of
a few lines of code.

MR: !78