libxfce4panel: Destroy plugin prefs dialog on plugin "destroy" signal
This avoids a crash when closing the panel via `xfce4-panel -q` (or when
removing an external plugin) while a plugin preferences dialog is open,
for which widget properties are binded to those of the plugin. This is
the case for all plugins embedded in the panel, except tasklist (which
uses a child widget instead). Destroying the dialog via a weak reference
on the plugin, i.e. on `dispose()`, is obviously too late, given the way
GObject handles the release of weak references internally.