libxfce4panel: Allow to show about dialog when prefs dialog is shown
There seems to be no reason not to allow showing the about dialog when
the preferences dialog is shown, especially since it is always possible
to show the preferences dialog when the about dialog is shown.

This condition was added a long time ago in
f825f21b6107eba51f3e64cf1034822a416fe348, without any particular
justification. Maybe a copy-paste error.

Fixes: #221