panel: Silence allocation warning for external plugins (V2)
This strives to impose a minimum size on external plugins only as long
as necessary to silence allocation warnings, thus avoiding side effects
noted in https://gitlab.xfce.org/xfce/xfce4-panel/-/merge_requests/84#note_51278.

There does not seem to be a reliable upper bound on the time interval
during which this minimum size is required, hence the use of a timeout
which is restarted as long as one of the external plugins requires a
reallocation on its own (this is typically needed at panel startup, when
several plugins are loading in parallel).

This should eliminate most allocation warnings (unless you are on a
really slow or overloaded system), without impacting the size of the
plugins beyond loading.

Fixes: #590
MR: !89
Related: !84