pager: Synchronize workspace layout between plugin instances
For internal Libwnck reasons, the workspace layout must be common to the
different instances of the plugin. So the best we can do is to
synchronize them properly, as much as possible. The number of rows
synchronizes well, but the orientation does not, which is why it is left
as it is in this commit.

Closes: #7
Related: !82, f1c08ac79d513d16b6e8bd0577be79d158fe80a7
See-also: https://bugzilla.gnome.org/show_bug.cgi?id=635341