libxfce4panel: Report a highlighted arrow button as blinking
This is contrary to what the documentation of the function said before,
but until 0ede27aeba2456d08f2c7204ce470b3014edfb79 (4.15.5) and since
ccd848851c5a207c999f83107a5365c5de8edd22 (4.9.0) the blinking timeout
has always been infinite. So in practice this will only change the
behavior of the function since 4.15.5, and will make it rather
consistent with what it was before. It's also probably more in line with
what you'd expect: the highlight at the end of the timeout is simply a
kind of "locked blinking".

Besides, this fixes a regression introduced by the above commit in
tasklist, when a window demands attention from another workspace, and
the blinking timeout is over: clicking on the window button does not
change the workspace anymore.

Related: !10