panel: Switch to pixels for the panel length in the prefs dialog
Internally, this length remains a proportion between 0 and 1, albeit
with increased precision. In Xfconf, it remains a percentage but again
with increased precision, so stored as a double. This ensures backward
compatibility.

So the user can still enter a percentage via Xfconf if he wants to. It
does not seem necessary to make the preferences dialog more cumbersome
by maintaining this possibility at this stage.

Fixes: #99