systray: Add debug info when item auto-terminates
It is not clear that this auto-termination is appropriate under any
circumstances. Perhaps one should either accept the failure or try to
re-register the item via the SnWatcher. It is likely that this is the
cause of unexpected item vanishing, as noted in #102, #243 and #599.
This commit should help debugging the problem.

Helps: #102, #243, #599