systray: Do not auto-terminate on error in a running proxy callback
These errors can have multiple causes, but if the proxy is still alive,
they are probably only temporary. If the proxy is dead or dying, the
item finalization should be triggered elsewhere (normally by
unregistering the status notifier).

The display of debug info (and thus the use of a GError) is kept in
`sn_item_get_all_properties_result()`, to make debugging easier if it
should have any side effects.

Helps: #102, #243, #599