Guard X11-specific/Wayland-incompatible code
This includes changes that allow the panel to start on Wayland without
crashing, with its Wayland-compatible embedded plugins (in the sense
that they don't crash the panel). The other embedded plugins are
automatically disabled.

The panel can still crash because of non-embedded plugins, because they
are now run as internal, i.e. in the same process as the panel. See the
changes to `panel-module.c` to add non-embedded plugins to the automatic
exclusion list.

Features disabled:
* Plugin dnd (xfce-panel-plugin.c)
* WM check at startup (panel-application.c)
* Running plugins as external (panel-module.c)
* Span monitors (panel-preferences-dialog.c, panel-window.c)
* Fix for #227 "Ensure that handles are always grabbable"
  (panel-window.c)
* Intellihide (panel-window.c)