wayland: panel: Ability to move the panel
Since there is no notion of global coordinates on Wayland,
`gtk_window_move()` does not work. Also, once initialized layer-shell,
the panel becomes static and it is no longer possible to use
`gtk_window_begin_move_drag()`, whose unconstrained movement type is not
suitable for the panel anyway (just a fallback if layer-shell is not
supported).

So we play with layer-shell anchors and margins to reproduce the panel
movement as it occurs on X11. At first it is enough to use the top and
left anchors, because we don't care about exclusive zones (X11 struts)
yet.

There is no need to manually grab the pointer anymore, as it is implied
on Wayland when a surface is clicked.

Known regressions:
* The grab is lost when changing monitor in automatic mode. Keeping the
  manual grab as on X11 is not a solution, because it prevents the
  automatic Wayland grab from working properly.