wayland: Set exclusive zones (X11 struts)
For exclusive zones to work properly, they must be set only on one edge,
or on one edge and its two perpendicular edges. Since we need at least
two edges to manage the movement of the panel, we are left with the
second option.

Also, all the margins must now be set so that the panel is the right
size and its movement proceeds correctly.