panel: Always include borders to compute snap position
If a motion event is sent twice in a row for the same pointer position
around `SNAP_DISTANCE` (this can happen if the pointer moves slowly),
the computation returns a snap position once, and `SNAP_POSITION_NONE`
the other time, because the panel has lost a pixel in between. This
makes the panel blink or bounce against the edge of the screen. Easier
to reproduce on Wayland.