wayland: panel: Autohide adaptations (foreign-toplevel-management part)
This partially restores the intellihide feature by hiding the panel if
the active window is maximized. For the rest, it would be necessary to
have access to the geometry of the foreign toplevel, which the protocol
does not allow in its current state, and which does not seem to be
possible with a reasonable effort by any other means.