clock: Implemented layout property for digital clock
A new property "digital-layout" is added to the digital clock to
store the layout selected by the user. Also in the properties dialog
box, now if the user selects "date" layout, then only the settings
related to date are shown. Similarly if "time" is selected then only
those settings are shown. And if "date, then time" or "time, then date"
is selected then both the settings are shown.