clock: Added date and time label to digital clock
Currently, the digital clock has a class of GtkLabel. So it can only
have one label and not two different labels for date and time. Now
the digital clock class is modified to GtkBox and it has two GtkLabel
"date_label" and "time_label" to show date and time respectively.