clock: Improve backward compatibility
Removes all the old code related to the no longer used "digital-format"
property, and queries Xfconf instead. If none of the new properties are
set and the old one exists, we use it as the time format, with a layout
that reproduces as much as possible the old rendering.

Helps: #641
Related: !101