Fix previous commits about menu positioning
Avoid iterating the main loop by hand. This is bad practice that may
change the order in which things should normally happen and lead to bad
surprises.

For popup_menu() use g_idle_add() instead. For position_widget(), just
delete this code since it was broken before and probably never called,
and put a hint in the doc instead.

Fixes: 7d0976db, 42b6032e